Convert QueryExecutionInfo to interface
- Convert `QueryExecutionInfo` to interface
- Add `DefaultQueryExecutionInfo` class in `callback` package
